+++ fe1-web/ingestion/handlers/index.ts
@@ -1,8 +1,8 @@
 export * from './Default';
-export * from './Lao';
-export * from './Meeting';
-export * from './RollCall';
-export * from './Witness';
+export * from './LaoHandler';
+export * from './MeetingHandler';
+export * from './RollCallHandler';
+export * from './WitnessHandler';
 export * from './ChirpHandler';
 export * from './ReactionHandler';
-export * from './Election';
+export * from './ElectionHandler';
